Make it render again
