+++ app/globals.css
@@ -1,0 +1,84 @@
+/* @tailwind base;
+@tailwind components;
+@tailwind utilities; */
+
+  body,
+  ul {
+    padding: 0;
+    margin: 0;
+  }
+  body {
+    font: 1em/1.4 "Fira Mono", "Menlo", monospace;
+    background: #fff;
+    color: #2a2a2a;
+    text-align: center;
+  }
+  a {
+    color: #333;
+    font-weight: bold;
+    text-decoration: underline;
+  }
+  h1 {
+    margin: 0 0 0.25em;
+    font-size: 2em;
+    font-weight: bold;
+  }
+  ul {
+    background: #2a2a2a;
+    color: #fff;
+  }
+  li {
+    position: absolute;
+    list-style: none;
+    border-top: 1px solid #fff;
+    width: 100%;
+    z-index: 1;
+    opacity: 0.1;
+  }
+  li:hover {
+    z-index: 100000;
+    opacity: 1;
+  }
+  li:hover span {
+    background: #fff;
+    color: #2a2a2a;
+  }
+  li span {
+    background: #2a2a2a;
+  }
+  li.current {
+    border-top: 1px solid #ff0000;
+    z-index: 100000;
+    opacity: 1;
+  }
+  li.current:hover span {
+    background: #ff0000;
+    color: #fff;
+  }
+  li.current span {
+    background: #ff0000;
+  }
+  header {
+    padding: 1em;
+    position: relative;
+    z-index: 2;
+    background: #fff;
+  }
+  header p {
+    text-align: left;
+    max-width: 55em;
+    margin: 0 auto 0.5em;
+  }
+  header p span {
+    color: #ff0000;
+  }
+  footer {
+    font-size: 80%;
+    padding: 1em;
+    max-width: 55em;
+    margin: 0 auto;
+  }
+  footer p {
+    text-align: left;
+  }
+
