+++ app/layout.tsx
@@ -1,16 +1,32 @@
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
-}
+import { Analytics } from "@vercel/analytics/react";
+import { Fira_Mono } from "next/font/google";
+import StyledJsxRegistry from "./registry";
+import type { Metadata } from "next";
+import "./globals.css";
+
+const fira = Fira_Mono({
+  weight: ["400", "700"],
+  subsets: ["latin"],
+});
+
+export const metadata: Metadata = {
+  title: "I am the fold",
+  description:
+    "An experiment to show how designing for The Fold can be treacherous",
+  viewport: "width=device-width initial-scale=1",
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
-    <html lang="en">
-      <body>{children}</body>
+    <html lang="en" className={fira.className}>
+      <body>
+        <StyledJsxRegistry>{children}</StyledJsxRegistry>
+      </body>
+      <Analytics />
     </html>
-  )
+  );
 }
